refactor(types): move MultiSearch types to Types.ts
- Move all MultiSearch type definitions from `MultiSearch.ts` to `Types.ts`
- Update imports in `MultiSearch.ts` to reference the moved types
- Re-export types from `MultiSearch.ts` to maintain API compatibility
- Remove redundant duplicate type definitions in `Types.ts`
- Update `RequestParams` interface to support MultiSearch stream configs
